requirements.txt
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.3
django-allauth==0.39.1
django-filter==2.1.0
django-filters==0.2.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djoser==1.7.0
gunicorn==19.9.0
httpie==1.0.2
idna==2.7
Markdown==3.1.1
oauthlib==3.0.1
pipenv==2018.11.26
psycopg2==2.8.3
Pygments==2.4.2
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.19.1
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.23
virtualenv==16.6.1
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
whitenoise==4.1.2
